pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automated-literature-review"
version = "0.1.0"
description = "Production-ready automated literature review pipeline for OpenAlex datasets."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Research Automation Team" }
]
dependencies = [
  "aiohttp>=3.9.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
